Fixed bug in DumpFrame on CheckerFlicker
CheckerFlicker used to behave incorrectly if dumping a frame in the past,
this is becuase it is a delayed init plugin.  As a workaround, if DumpFrame
detects a frame in the past, it stops and starts the plugin, waiting for it to be fully initted.

Modified Stop.m and Start.m to wait for the plugin